refactor: 파일 용도별 분리
