+++ src/components/Navigate.js
@@ -1,0 +1,25 @@
+import { state } from "../state";
+
+const Navigate = () => {
+  if (state.user) {
+    return `
+      <nav class="bg-white shadow-md p-2 sticky top-14">
+        <ul class="flex justify-around">
+          <li><a href="/" class="text-blue-600">홈</a></li>
+          <li><a href="/profile" class="text-gray-600">프로필</a></li>
+          <li><a href="/logout" class="text-gray-600">로그아웃</a></li>
+        </ul>
+      </nav>
+`;
+  }
+  return `
+      <nav class="bg-white shadow-md p-2 sticky top-14">
+        <ul class="flex justify-around">
+          <li><a href="/" class="text-blue-600">홈</a></li>
+          <li><a href="/login" class="text-gray-600">로그인</a></li>
+        </ul>
+      </nav>
+`;
+};
+
+export default Navigate;
