+++ src/constant.js
@@ -1,0 +1,6 @@
+export const userInfo = {
+  username: "testuser",
+  email: "test@example.com",
+  bio: "hi",
+  password: "1234",
+};
