+++ src/state/index.js
@@ -1,0 +1,44 @@
+import { render } from "../routes";
+
+const state = {
+  user: JSON.parse(localStorage.getItem("user")) || null,
+  posts: [
+    {
+      name: "홍길동",
+      time: "5분 전",
+      content: "오늘 날씨가 정말 좋네요. 다들 좋은 하루 보내세요!",
+    },
+    {
+      name: "김철수",
+      time: "15분 전",
+      content: "새로운 프로젝트를 시작했어요. 열심히 코딩 중입니다!",
+    },
+    {
+      name: "이영희",
+      time: "30분 전",
+      content: "오늘 점심 메뉴 추천 받습니다. 뭐가 좋을까요?",
+    },
+    {
+      name: "박민수",
+      time: "1시간 전",
+      content: "주말에 등산 가실 분 계신가요? 함께 가요!",
+    },
+    {
+      name: "정수연",
+      time: "2시간 전",
+      content: "새로 나온 영화 재미있대요. 같이 보러 갈 사람?",
+    },
+  ],
+};
+
+const initUser = () => {
+  state.user = JSON.parse(localStorage.getItem("user")) || null;
+};
+
+const setUser = (user) => {
+  state.user = user;
+  localStorage.setItem("user", JSON.stringify(user));
+  render(window.location.pathname);
+};
+
+export { state, initUser, setUser };
